--- a/.github/workflows/publish.yml
+++ b/.github/workflows/publish.yml
@@ -9,7 +9,7 @@ jobs:
   build:
     env:
       # Nuget package version. Change it to publish a new version.
-      version: 1.0.3
+      version: 1.0.4
 
     name: Publish NuGet Package