--- a/LICENSE
+++ b/LICENSE
@@ -1,6 +1,6 @@
 BSD 3-Clause License
 
-Copyright (c) 2020, Anton Gogolev
+Copyright (c) 2020, Anton Gogolev and contributors
 All rights reserved.
 
 Redistribution and use in source and binary forms, with or without