--- a/src/NLog.Loki.Tests/NLog.Loki.Tests.csproj
+++ b/src/NLog.Loki.Tests/NLog.Loki.Tests.csproj
@@ -1,7 +1,7 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
+    <TargetFramework>net6.0</TargetFramework>
 
     <IsPackable>false</IsPackable>
   </PropertyGroup>