--- a/src/NLog.Loki.sln
+++ b/src/NLog.Loki.sln
@@ -1,7 +1,7 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.29806.167
+# Visual Studio Version 17
+VisualStudioVersion = 17.0.31919.166
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NLog.Loki", "NLog.Loki\NLog.Loki.csproj", "{E47CC9DB-0B43-48A7-BAA1-A03DB0049D76}"
 EndProject
@@ -12,6 +12,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 	EndProjectSection
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NLog.Loki.Tests", "NLog.Loki.Tests\NLog.Loki.Tests.csproj", "{901267E9-BE8E-4FBF-AF90-D5BF5FB6F70D}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Benchmark", "Benchmark\Benchmark.csproj", "{0880E1C9-7C30-4EA3-AE34-0CC02B22C628}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -27,6 +29,10 @@ Global
 		{901267E9-BE8E-4FBF-AF90-D5BF5FB6F70D}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{901267E9-BE8E-4FBF-AF90-D5BF5FB6F70D}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{901267E9-BE8E-4FBF-AF90-D5BF5FB6F70D}.Release|Any CPU.Build.0 = Release|Any CPU
+		{0880E1C9-7C30-4EA3-AE34-0CC02B22C628}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{0880E1C9-7C30-4EA3-AE34-0CC02B22C628}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{0880E1C9-7C30-4EA3-AE34-0CC02B22C628}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{0880E1C9-7C30-4EA3-AE34-0CC02B22C628}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE