--- a/src/NLog.Loki/AssemblyInfo.cs
+++ b/src/NLog.Loki/AssemblyInfo.cs
@@ -1,3 +1,4 @@
 using System.Runtime.CompilerServices;
 
-[assembly:InternalsVisibleTo("NLog.Loki.Tests")]
+[assembly: InternalsVisibleTo("NLog.Loki.Tests")]
+[assembly: InternalsVisibleTo("Benchmark")]