--- a/src/NLog.Loki/Impl/IJsonSerializer.cs
+++ b/src/NLog.Loki/Impl/IJsonSerializer.cs
@@ -1,9 +1,11 @@
+using System.Collections.Generic;
 using System.Threading.Tasks;
+using NLog.Loki.Model;
 
 namespace NLog.Loki.Impl
 {
     internal interface IJsonSerializer
     {
-        Task SerializeAsync(object instance, JsonTextWriter jsonTextWriter);
+        Task SerializeAsync(IEnumerable<LokiEvent> instance, JsonTextWriter jsonTextWriter);
     }
 }