--- a/src/NLog.Loki/LokiHttpClient.cs
+++ b/src/NLog.Loki/LokiHttpClient.cs
@@ -12,9 +12,9 @@ namespace NLog.Loki
             this.httpClient = httpClient;
         }
 
-        public async Task<HttpResponseMessage> PostAsync(string requestUri, HttpContent httpContent)
+        public Task<HttpResponseMessage> PostAsync(string requestUri, HttpContent httpContent)
         {
-            return await httpClient.PostAsync(requestUri, httpContent);
+            return httpClient.PostAsync(requestUri, httpContent);
         }
     }
 }