--- a/src/NLog.Loki/LokiHttpClient.cs
+++ b/src/NLog.Loki/LokiHttpClient.cs
@@ -3,7 +3,7 @@ using System.Threading.Tasks;
 
 namespace NLog.Loki
 {
-    public class LokiHttpClient : ILokiHttpClient
+    internal class LokiHttpClient : ILokiHttpClient
     {
         private readonly HttpClient httpClient;