--- a/src/NLog.Loki/LokiTarget.cs
+++ b/src/NLog.Loki/LokiTarget.cs
@@ -82,7 +82,7 @@ namespace NLog.Loki
                 if(uri.Scheme == Uri.UriSchemeHttp || uri.Scheme == Uri.UriSchemeHttps)
                 {
                     var lokiHttpClient = LokiHttpClientFactory(uri);
-                    var httpLokiTransport = new HttpLokiTransport(uri, lokiHttpClient);
+                    var httpLokiTransport = new HttpLokiTransport(lokiHttpClient);
 
                     return httpLokiTransport;
                 }