--- a/src/NLog.Loki/NLog.Loki.csproj
+++ b/src/NLog.Loki/NLog.Loki.csproj
@@ -14,6 +14,8 @@
     <FileVersion>0.0.0.0</FileVersion>
     <PackageLicenseFile>LICENSE</PackageLicenseFile>
     <PackageReadmeFile>README.md</PackageReadmeFile>
+    <IncludeSymbols>true</IncludeSymbols>
+    <SymbolPackageFormat>snupkg</SymbolPackageFormat>
   </PropertyGroup>
 
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Release|AnyCPU'">