… .github/workflows/ci.yml
… .semver
… README.md
… src/Benchmark/Benchmark.csproj
… src/Benchmark/Benchmarks.cs
… src/Benchmark/Program.cs
… src/Benchmark/Transport.cs
… src/Benchmark/docker-compose.yml
… src/Benchmark/loki-config.yml
… src/Benchmark/nlog.config
… src/NLog.Loki.Tests/Impl/JsonTextWriterExtensionsTests.cs
… src/NLog.Loki.Tests/Impl/JsonTextWriterTests.cs
… src/NLog.Loki/Impl/JsonTextWriter.cs
… src/NLog.Loki/Impl/JsonTextWriterExtensions.cs
… src/NLog.Loki/Impl/TypeExtensions.cs
… src/NLog.Loki/NullLokiTransport.cs